Fix BadWindow errors that occurred when closing an X11 window.
This commit is based on OpenTK's fix in:
https://github.com/opentk/opentk/commit/ec31675fc87883cb81c4df3498385a7fe54eae09

...and was adapted for ClassicalSharp's modified OpenTK code.

This also fixes options not being saved on Linux when the client or
launcher closes, at least on Debian Jessie. The default error handler
for Xlib is to print a message explaining the error to the console and
then call exit(1). If that happens, the Options.Save() calls in
Game.Dispose() and LauncherWindow.Run() are never reached.