Made inventory switching re-bindable and fixed CRTL selecting in inventory screen closing inventory screen.
Inventory switching(I.E. Pressing ALT in-game showing a second inventory) is now re-bindable.

Pressing the CRTL keys should've allowed the user to select a block without the inventory screen closing, but this was not the case. Instead, that functionality was binded to the ALT keys. It has now been re-binded to the CRTL keys.