Offset vertices instead of indices in the OpenGL backend.
This fixes issues when using mesa and the i915 dri backend (tested on lubuntu, may affect other distributions), specifically for the 945GM and related gpus. This fixes issues such as parts of the map being invisible, blocks in the inventory being invisible depending on the number of sprite and non-sprite blocks in the inventory, position text becoming invisible depending on your location in the world, axis lines sometimes being invisible, entity models sometimes being invisible, particles sometimes being visibile or invisible depending on the frame, which had different invisibility behaviour every time the client was run.