Fix Launcher_MarkDirty accidentally producing dirty rectanglesthat extended outside of the window.
Normally the consequence of this was redrawing a bit more than needed because width was > height. This wouldn't cause a crash with the current OS 'draw framebuffer' functions because they just clip the region outside the window, but could cause crashes when writing custom Window_DrawFramebuffer implementations and wasn't aware of this.