Fix rare case on non-Windows where Waitable_Wait would miss a signal
pthread_cond_signal is documented to have no effect if no threads are blocked/waiting. As such, if another thread called pthread_cond_wait afterwards, it would not realise it was supposed to be signalled already, and then incorrectly proceed to block while waiting for another signal. This issue manifested in the http worker thread, which would rarely cause the launcher to get stuck 'checking for updates' until some other http request got queued, which would wake up the stuck worker thread. Fix this by adding a field that keeps track of whether the waitable is currently in the 'signalled' state.