If an attempt is made to extract texture pack while context is lost, add code to reload texture pack when context is recreated
The most noticable way this issue manifests is when you try to start the game (direct3d9) while a fullscreen application is already running. This causes creating the device to initially fail, but when the context is later recreated (because you toggled out of the fullscreen application), all the textures would appear white because the initial texture pack had attempted to load while context was initially lost.

Of course, the game still crashes when you try to resize the window next (or something else that causes context to get lost), but progress still.