Remove dividing .cw spawn by 32 when spawn is outside map backwards compatibility patch
This means if your position is outside the map and you save the map, when you load the map, you appear at this position again, instead of the position divided by 32.

This was a workaround for ClassicalSharp maps saved before 0.98.6. Considering that was relased on Feb 2016, I think it's an acceptable loss that a few maps from waayyy back then have the wrong spawn when loaded.