Webclient: Space should always have its default action prevented unless text input keyboard is 'open'
This change is necessary because Safari scrolls the page when you press space and the default action is not prevented. However, we also can't just always prevent the default action, because then the key press event for ' ' would never get raised for text input. So this seems like a reasonable compromise.