Webclient: WIP on properly supporting touchscreen input
When touch mode is off and a touchstart event is received, switch to touch mode. When touch mode is on and a mousedown event is received, disable touch mode. Unfortunately this still has a problem where touch mode will always get disabled if you touch while the keyboard is open. (because the game does NOT prevent the default behaviour of touchstart becoming a mousedown event when the keyboard is open, so that you can still click on the caret in text input box)