Add hold model
I know this is a bit of a hacky work-around but I think it could be beneficial to allow players to be able to see blocks they're holding (will make an MCGalaxy plugin later in support of this). Could be useful for the time being until (if) something like #153 is implemented.