Properly handle poll/select for writing
The last error code should be checked when select/poll returns true for writability, not false like it was before.

This should e.g. fix connecting to a non-existent LAN server showing 'disconnected from server' error instead of 'failed to connect to server' error. This was because since the async connect call was wrongly treated as having succeeded, the send method was then called, which immediately failed.